Timers and Message Queues

+++ RTXOff/rtxoff_thread.cpp
@@ -1136,7 +1136,7 @@ uint32_t osThreadFlagsWait (uint32_t flags, uint32_t options, uint32_t timeout)
 
 			if(thread->waitValPresent)
 			{
-				thread_flags = thread->waitExitVal;
+				thread_flags = static_cast<uint32_t>(thread->waitExitVal);
 				thread->waitValPresent = false;
 			}
 			else
